backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=2.1.1",
    "python-dotenv>=1.1.0",
]
